backend/pyproject.toml
[project]
name = "lr-autotag-backend"
version = "0.1.0"
description = "Backend service for LR-AutoTag — automatic photo keywording for Lightroom Classic"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "pydantic-settings>=2.4",
    "psycopg[binary,pool]>=3.2",
    "httpx>=0.27",
    "Pillow>=10.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "ruff>=0.6",
]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "unit: Unit tests — no external dependencies",
    "integration: Integration tests — requires PostgreSQL",
    "system: System tests — full FastAPI stack",
    "performance: Performance/benchmark tests",
    "security: Security-focused tests",
    "slow: Tests taking >5s",
]
